MEGA PUSH

+++ core/src/com/mygdx/clashofclans/Tokens/Defenses/Bomb.java
@@ -11,7 +11,7 @@ public class Bomb extends Defense {
 
     Animation<TextureRegion> bombAnimation = GifDecoder.loadGIFAnimation(Animation.PlayMode.LOOP, Gdx.files.internal(BOMB_DEFENSE_ANIMATION).read());
 
-    public Bomb(int pInitialX, int pInitialY) {
+    public Bomb(float pInitialX, float pInitialY) {
         super(pInitialX, pInitialY, BOMB_DEFENSE_LIFE, BOMB_DEFENSE_RANGE, BOMB_DEFENSE_ATTACKRATE);
     }
